requirements.txt
Flask==2.2.3
Werkzeug==2.3.7
SQLAlchemy==2.0.7
PyExifTool==0.5.5
arrow==1.2.3
python-magic==0.4.27
pyxdg==0.28
setuptools==65.7.0
voluptuous==0.13.1
wheel==0.38.4
